+++ src/hooks/useCountdown.js
@@ -1,3 +1,34 @@
-export const useCountdown = (targetDate) => {
+import { useState, useEffect } from "react";
+
+export const useCountdown = () => {
+  const targetDate = new Date("2025-08-25T00:00:00");
+
+  const calculateTimeLeft = (targetDate) => {
+    const difference = targetDate.getTime() - new Date().getTime();
+
+    if (difference <= 0) {
+      return { days: 0, hours: 0, minutes: 0, seconds: 0 };
+    }
+
+    const days = Math.floor(difference / (1000 * 60 * 60 * 24));
+    const hours = Math.floor((difference / (1000 * 60 * 60)) % 24);
+    const minutes = Math.floor((difference / (1000 * 60)) % 60);
+    const seconds = Math.floor((difference / 1000) % 60);
+
+    return { days: days,
+      hours : hours, 
+      minutes: minutes, 
+      seconds : seconds };
+  };
+  const [timeLeft, setTimeLeft] = useState(() => calculateTimeLeft(targetDate));
+
+  useEffect(() => {
+    const timer = setInterval(() => {
+      setTimeLeft(calculateTimeLeft(targetDate));
+    }, 1000);
+
+    return () => clearInterval(timer);
+  }, [targetDate]);
+
   return timeLeft;
 };
